POC: add discovery of PostgreSQL databases and push them as entities
We take all services with label application=spilo and try to connect to the
database cluster using .svc.cluster.local DNS name.

We assume that all Spilo clusters have the same user and password configured
with wich we can connect and list the database names.  The username and the
password need to be provided in two new environment variables, namely:
ZMON_AGENT_POSTGRES_{USER,PASS}.  We assume that we can reuse pre-existing
user that is configured for ZMON worker already.

For every non-system database we create a separate entity of type
'postgresql_database' and configure the 'shards' property so that ZMON checks
can use `sql()` function without further parameters to access the database.